+++ anthem/lyrics/modules.py
@@ -24,6 +24,17 @@ def update_translations(ctx, module_list):
         raise AnthemError(u"You have to provide a list of "
                           "module's name to update the translations")
 
+    ir_module = ctx.env['ir.module.module']
+    if hasattr(ir_module, 'update_translations'):
+        # Odoo version <= 10.0
+        method_name = 'update_translations'
+    else:
+        # Odoo version >= 11.0
+        method_name = '_update_translations'
+
     for module in module_list:
-        ctx.env['ir.module.module'].with_context(overwrite=True).search(
-            [('name', '=', module)]).update_translations()
+        getattr(
+            ir_module.with_context(overwrite=True).search(
+                [('name', '=', module)]),
+            method_name
+        )()